api/requirements.txt
flask
psycopg2-binary
pandas
numpy
flask-cors
sqlalchemy
requests
python-dotenv
gunicorn